university/tests/requirements.txt
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.25.1